move gc packages around
