+++ versioned/gc/core/src/main/java/org/projectnessie/versioned/gc/AssetKeyConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.versioned.gc.core;
+package org.projectnessie.versioned.gc;
 
 import java.util.stream.Stream;
