+++ versioned/gc/pom.xml
@@ -21,21 +21,16 @@
 
   <parent>
     <groupId>org.projectnessie</groupId>
-    <artifactId>nessie</artifactId>
+    <artifactId>nessie-versioned</artifactId>
     <version>0.3.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>nessie-versioned</artifactId>
+  <artifactId>nessie-versioned-gc-parent</artifactId>
   <packaging>pom</packaging>
 
-  <name>Nessie - Versioned - Parent</name>
+  <name>Nessie - Versioned - GC for Spark - Parent</name>
 
   <modules>
-    <module>jgit</module>
-    <module>memory</module>
-    <module>spi</module>
-    <module>tests</module>
-    <module>tiered</module>
-    <module>gc</module>
+    <module>core</module>
   </modules>
 </project>
